Use a seprate plist file for the user study configs
