+++ LLMonFHIR/Supporting Files/UserStudyConfig.plist
