+++ LLMonFHIR/Supporting Files/UserStudyConfig.plist.license
@@ -1,0 +1,6 @@
+
+This source file is part of the Stanford LLM on FHIR project
+
+SPDX-FileCopyrightText: 2025 Stanford University
+
+SPDX-License-Identifier: MIT
